Add support for mobile phones

+++ frontend/src/HeartbeatManager.js
@@ -4,18 +4,23 @@ module.exports = function(socket, chat, userDao) {
 	var _ = require('underscore');
 
 	var lastHeartbeat = {}
+	var mobileUser = {};
 
 	that.start = function() {
 		setInterval(function() {
 			var current = new Date().getTime();
 			_.each(lastHeartbeat, function(data, username) {
+				if (mobileUser[username]) {
+					// Mobile users will not be timed out.
+					return;
+				}
 				var user = data.user;
 				var timestamp = data.timestamp;
 				if (current - 10000 > timestamp) {
 					delete lastHeartbeat[username];
 					userDao.removeUser(user);
 					chat.userTimedOut(user);
-				}	
+				}
 			});
 		}, 10000);
 	}
@@ -26,5 +31,9 @@ module.exports = function(socket, chat, userDao) {
 			user: user
 		}
 	});
-	
+
+	socket.on('mobileStatus', function(user, status) {
+		mobileUser[user.username] = status;
+	});
+
 }
